interface/requirements.txt
nicegui==1.4.29
pyserial==3.5